Actualización, usuario puede mantenerse logueado
