+++ client/src/App.tsx
@@ -1,6 +1,7 @@
 /// IMPORTS
 import { Routes, Route, useLocation } from 'react-router-dom';  
 import { Toaster } from 'react-hot-toast'
+import { useEffect } from 'react';
 // VIEWS
 import Landing from './views/Landing/Landing';
 import Home from './views/Home/Home'
@@ -15,23 +16,34 @@ import Cart from './views/Cart/Cart';
 import AboutUs from './views/aboutUs/AboutUs';
 import Contact from './views/Contact/Contact';
 import Creation from './views/Creation/Creation';
-import { Toaster } from 'react-hot-toast'
 import Pay from './views/Pay/Pay';
 
 // COMPONENTS
 import NavBar from "./components/navbar/NavBar"
 import Footer from "./components/footer/Footer"
+import { verificationLogin } from './redux/actions/actions';
 // STYLES
 import './App.css';
 import SuccessPay from './views/Pay/succesPay';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { AppState } from './redux/reducer';
 
 // APP
 function App() {
+  const dispatch = useDispatch()
   const location = useLocation();
   const hasPermissions = false;
   const { accessLogin } = useSelector((state: AppState) => state) 
+  useEffect(()=>{
+    const userJSON = localStorage.getItem("user")
+    console.log(userJSON);
+    
+    if(userJSON){
+      const user = JSON.parse(userJSON)
+      dispatch(verificationLogin(user))
+      console.log(user);
+    }
+  }, [dispatch])
   console.log(accessLogin)
   return (
     <div>
