+++ client/src/main.tsx
@@ -7,8 +7,8 @@ import { Provider } from 'react-redux'
 import { store } from './redux/store';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import axios from 'axios';
-// axios.defaults.baseURL = 'http://localhost:3001'
-axios.defaults.baseURL = 'https://craftbeer.up.railway.app/'
+axios.defaults.baseURL = 'http://localhost:3001'
+//axios.defaults.baseURL = 'https://craftbeer.up.railway.app/'
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
